requirements.txt
mysql-connector-python
getpass